Preserve uri parameters on indialog-requests
Some devices (e.g. Snom) send a Contact like this:

sip:user@ip:port;line=abcde

With the current approach splitting the uri to parts, the line
parameter is dropped, and re-INVITE or BYE are rejected by those
devices due to missing line parameters.

The new approach is to take the uri as-is and slightly modify the
uri extraction regex, extracting the full uri including all params
and use that as r-uri.