Do not run dispatcher callback every second but only if needed
Method queue_expire is already prepared to return time to next expire or
undef is nothing is waiting, but timer callback does not use it and
instruct loop to run this callback every one second. Fix dispatcher code to
setup and cancel timer when needed.

This change radically decrease usage of processor time as Net::SIP's
Dispatcher does not have to be called every second if there is nothing to
dispatch and therefore operating system does not have to interrupt Net::SIP
application from select() syscall.