Fix setting IP address in Contact and Via headers when leg is listening on wildcard address
In some situation it makes sense to create Net::SIP::Leg to listen on
wildcard IPv4 address 0.0.0.0 or IPv6 ::

But in this case it is needed to fill correct IP address in Contact and Via
headers of outgoing interface.

This change fixes IP address in Contact and Via headers via laddr4dst()
function which retrieves local IP address uses for connecting to
destination address.